исправил бота для cmd, сделал javadoc для интерфейса Connector
