+++ src/main/java/org/matmech/connector/Connector.java
@@ -1,11 +1,8 @@
 package org.matmech.connector;
 
+/**
+ * Общий интерфейс для всех ботов с обязательным методом <i>start</i>, который запускает всего бота
+ */
 public interface Connector {
     void start();
-    default void postRequest() {
-        System.out.println("Post request!");
-    }
-    default void getRequest() {
-        System.out.println("Get request!");
-    }
 }
